+++ mocks/mockRai.js
