--- ogTestsAndSamples/sample/sample.css
@@ -1,105 +1,0 @@
-.counters {
-	margin-bottom: 15px;
-}
-
-.password {
-	margin-top: 45px;
-}
-
-.accounts {
-	min-width: 400px;
-	margin-top: 30px;
-	margin-bottom: 75px;
-}
-
-.line {
-	width: 700px;
-	height: 15px;
-}
-
-.account {
-	float:left;
-	white-space: nowrap;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	width: 510px;
-}
-
-@media (max-width: 700px) {
-	.line { width: 400px; }
-	.account{ width: 250px; }
-}
-
-.balance {
-	float:right;
-	width: 150px;
-	text-align: right;
-}
-
-.remove {
-	float:right;
-	width: 40px;
-	text-align: right;
-	color: red;
-	font-weight: bold;
-	cursor: pointer;
-}
-.remove:hover {
-	text-decoration: underline;
-}
-
-.head {
-	font-weight: bold;
-	text-align: center
-}
-
-.history {
-	
-}
-
-.hist_hash , .hist_account{
-	white-space: nowrap;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	max-width: 200px;
-}
-
-.hist_amount{
-	text-align: right;
-}
-
-.hide {
-	display: none;
-}
-
-.button {
-	background: #3498db;
-	background-image: -webkit-linear-gradient(top, #3498db, #2980b9);
-	background-image: -moz-linear-gradient(top, #3498db, #2980b9);
-	background-image: -ms-linear-gradient(top, #3498db, #2980b9);
-	background-image: -o-linear-gradient(top, #3498db, #2980b9);
-	background-image: linear-gradient(to bottom, #3498db, #2980b9);
-	-webkit-border-radius: 6;
-	-moz-border-radius: 6;
-	border-radius: 6px;
-	font-family: Arial;
-	color: #ffffff;
-	font-size: 14px;
-	padding: 4px 8px 4px 8px;
-	text-decoration: none;
-	display: inline-block;
-}
-
-.button:hover {
-	background: #3cb0fd;
-	background-image: -webkit-linear-gradient(top, #3cb0fd, #3498db);
-	background-image: -moz-linear-gradient(top, #3cb0fd, #3498db);
-	background-image: -ms-linear-gradient(top, #3cb0fd, #3498db);
-	background-image: -o-linear-gradient(top, #3cb0fd, #3498db);
-	background-image: linear-gradient(to bottom, #3cb0fd, #3498db);
-	text-decoration: none;
-}
-
-input.readonly {
-	background-color: #cccccc;
-}
