--- ogTestsAndSamples/scripts/transaction_booster.html
@@ -1,105 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>RaiBlocksJS Transaction Booster</title>
-<meta charset="utf-8">
-
-<link rel="stylesheet" media="all" href="..\sample.css">
-<script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
-<!-- JS BigNumber library https://github.com/MikeMcl/bignumber.js -->
-<script type="text/javascript" src="..\bignumber.min.js"></script>
-<script type="text/javascript" src="..\rai.rpc.js"></script>
-<script type="text/javascript" src="..\rai.extended.js"></script>
-<script type="text/javascript" src="..\rai.community.js"></script>
-
-</head>
-<body>
-
-<!-- General page scripts -->
-<script type="text/javascript">
-	json_load = function(url, data) {
-		try {
-				let xhr;
-				xhr = new XMLHttpRequest();
-				xhr.open("POST", url, false);
-				
-				xhr.send(data);
-				
-				if (xhr.readyState == 4 && xhr.status == 200) {
-					let json = JSON.parse(xhr.responseText);
-					return json;
-			}
-		}
-		catch (ex) {
-			alert(ex);
-			console.error(ex.message);
-		}
-	}
-
-
-	transaction_republisher = function(wallet, host) {
-		if (typeof host == 'undefined') { host = 'http://localhost:7076'; } // if not initialized
-		var rai = new Rai(host);
-		var frontiers = json_load("https://raiwallet.info/frontiers.json", JSON.stringify({}));
-		var wallet_frontiers = rai.wallet_frontiers(wallet);
-		
-		var accounts_list = rai.account_list(wallet);
-		
-		$.each(accounts_list, function(){
-			let frontier = wallet_frontiers[this];
-			if (typeof frontier != 'undefined') {
-				let chain_length = 128;
-				let chain = rai.chain(frontier, chain_length);
-				let ext_frontier = frontiers['frontiers'][this];
-				if (typeof ext_frontier == 'undefined') { // Empty external
-					console.log('Frontier not found');
-					alert("Rebroadcast block: " + chain[chain.length - 1]);
-					try {
-						rai.republish(chain[chain.length - 1]);
-					}
-					catch (ex) {
-						// do nothing
-					}
-				}
-				else {
-					for (let i = 0; i < chain.length; i++) {
-						current_hash = chain[i];
-						if (chain[0] == ext_frontier) break;
-						else if (current_hash == ext_frontier) {
-							console.log(i);
-							alert("Rebroadcast block: " + current_hash);
-							try {
-								rai.republish(current_hash);
-							}
-							catch (ex) {
-								// do nothing
-							}
-						}
-						else if ((i + 1) == chain.length) {
-							console.log(i);
-							alert("Rebroadcast block: " + chain[chain.length - 1]);
-							try {
-								rai.republish(chain[chain.length - 1]); // Send everything, if not found
-							}
-							catch (ex) {
-								// do nothing
-							}
-						}
-					}
-				}
-			}
-		});
-	}
-
-
-var wallet = prompt( " Enter wallet from config.json "); // replace promt with your wallet actual id
-var host = "http://localhost:7076"; // default host
-var rai = new Rai(host); // connection
-transaction_republisher(wallet, host);
-transaction_republisher(wallet, host);
-transaction_republisher(wallet, host);
-rai.bootstrap_any(); // bootstrap
-</script>
-
-</body>
-</html>
