--- ogTestsAndSamples/tests/BlockHash.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    
-</head>
-
-<body>
-    <script src="../blake2b.js"></script>
-    <script src="../nacl.js"></script>
-    <script src="../rai.rpc.js"></script>
-    <script src="../rai.extended.js"></script>
-    <script src="../rai.pow.js"></script>
-    <script>
-        
-        var rai = new Rai();
-        
-        // https://raiblockscommunity.net/block/index.php?h=130E1FD0263DFA9881CC47BC51D391FA92BD2BA23CC1482394EA3B536BC7B004
-        // 130E1FD0263DFA9881CC47BC51D391FA92BD2BA23CC1482394EA3B536BC7B004
-        console.log(
-            rai.computeBlockHash('send', {
-                    previous: "8339D087F403C66BFF9817E5A2CF6215AB5EF69B200DD93CCB4676126BC65FB8",
-                    destination: rai.ext_account_key('xrb_118hywx9xfy8oc8tzgs9ba736tdkztnsjwqn6r71og4isn7tnhuqygrqsmjg'),
-                    balance: "0000000665EA1FD2BB39ED14DE000000"
-                }
-            )
-        );
-        
-        // https://raiblockscommunity.net/block/index.php?h=46AD46F2B3FD8C9F74E5B43B6FEEF34A4D1B39A2763BED4FEBC2EC7C57792238
-        // 46AD46F2B3FD8C9F74E5B43B6FEEF34A4D1B39A2763BED4FEBC2EC7C57792238	
-        console.log(
-            rai.computeBlockHash('receive', {
-                    previous: "8DEEC883A1E191E36F982289E4A5E1CD9BDA413B5A6FDE93E2F6D817AEAF2EF3",
-                    source: "1DB10F8F4752FCBE066C90185817C53D817B672CB1EF8755314EFD73701A9778"
-                }
-            )
-        );
-        
-        // https://raiblockscommunity.net/block/index.php?h=63C86899FE28B0A7DB3F0C11AF8CAEDF79896F5DFA07BA5E2B4F173C5A3E314A
-        // 63C86899FE28B0A7DB3F0C11AF8CAEDF79896F5DFA07BA5E2B4F173C5A3E314A
-        console.log(
-            rai.computeBlockHash('open', {
-                    source: "D6F390BF06C79E594459DB5031D0D12C1D03257E9A3585FE9F7ADEBF5CC9C0BC",
-                    representative: rai.ext_account_key('xrb_3arg3asgtigae3xckabaaewkx3bzsh7nwz7jkmjos79ihyaxwphhm6qgjps4'), 
-                    account: rai.ext_account_key('xrb_3dgrbyse8gusrk5b4xg6xnqxt5t69m67q6br1wmzjr76uzp5ax17smc48br6')
-                }
-            )
-        );
-        
-        // https://raiblockscommunity.net/block/index.php?h=FFCFCE8BD7104CCCB03FD8D723807BC755B9B6797368A68970B5490720EBB74F
-        // FFCFCE8BD7104CCCB03FD8D723807BC755B9B6797368A68970B5490720EBB74F
-        console.log(
-            rai.computeBlockHash('change', {
-                    previous: "C661E29551843F2CE75506DE3B8189CE0A40735CE3D4387C6B88B29758AD65F4",
-                    representative: rai.ext_account_key('xrb_1ma5dct7jdc8o45135xbr1bwbsixr47xugisu5fnwo69byhcno5u946smqfp')
-                }
-            )
-        );
-        
-    </script>
-    
-</body>
-
-</html>
