--- ogTestsAndSamples/tests/SeedTest.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>Seed test sample</title>
-<meta charset="utf-8">
-<link rel="stylesheet" media="all" href="..\sample.css">
-<script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
-<!-- JS Blake library https://github.com/dcposch/blakejs -->
-<script type="text/javascript" src="..\blake2b.js"></script>
-<script type="text/javascript" src="..\rai.rpc.js"></script>
-<script type="text/javascript" src="..\nacl.js"></script>
-<script type="text/javascript" src="..\rai.extended.js"></script>
-<style>
-div {
-	text-align: center;
-	padding: 10px;
-	font-size: 24px;
-}
-input {
-	width: 500px;
-}
-#seed_count {
-	width: 70px;
-}
-</style>
-
-</head>
-<body>
-
-<!-- Page content -->
-<div id="seed_input">
-	Enter seed to generate accounts
-	<form name="seed_input">
-		<input id="seed_input_text" value="0000000000000000000000000000000000000000000000000000000000000000">
-		<br />Count: <input id="seed_count" value="100">
-	</form>
-	<div id="start" class="button">Generate accounts</div>
-</div>
-
-<div id="threads"></div>
-<div id="hash" style="font-size: 22px;"></div>
-<div id="time"></div>
-<table id="accounts" align="center">
-<tr><th>ID</th><th>Account</th></tr>
-</table>
-
-<!-- General page scripts -->
-<script type="text/javascript">
-var acc_explorer = "https://raiblockscommunity.net/account/index.php?acc=";
-$( "#start" ).click(function() {
-	var seed = $('#seed_input_text')[0].value;
-	var count = parseInt($('#seed_count')[0].value);
-	$("tr").remove();
-	$("#accounts").append("<tr><th>ID</th><th>Account</th></tr>");
-	for (let i = 0; i < count; i++) {
-		let key = XRB.seed_key(seed, i);
-		let account = XRB.key_account(key);
-		$("#accounts").append("<tr><td>" + i + "</td><td><a href=\"" + acc_explorer + account + "\">" + account + "</a></td></tr>");
-		console.log(account);
-	}
-});
-</script>
-
-</body>
-</html>
